requirements.txt
asgiref==3.7.2
astroid==3.0.1
colorama==0.4.6
dill==0.3.7
Django==4.2.6
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-meta==2.4.0
django-modelcluster==6.1
django-robots==6.1
django-seo==0.3.5
isort==5.12.0
mccabe==0.7.0
Pillow==10.1.0
platformdirs==3.11.0
pylint==3.0.2
pylint-celery==0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
